// avantio-react-challenge/src/accommodationForm/values.enum.ts
export enum formValuesEnum {
  ACCOMMODATION_NAME = "accommodationName",
  ACCOMMODATION_ADDRESS = "accommodationAddress",
  ACCOMMODATION_DESCRIPTION = "accommodationDescription",
  ACCOMMODATION_TYPE = "accommodationType",
  ACCOMMODATION_IMAGES = "accommodationImages",
  OWNER_NAME = "ownerName",
  OWNER_EMAIL = "ownerEmail",
  OWNER_PHONE = "ownerPhone",
}

export enum stepsEnum {
  ACCOMODATION_STEP,
  OWNER_STEP,
  SUMMARY_STEP,
}

export enum accommodationTypesEnum {
  APARTMENT = "Apartment",
  VILLA = "Villa",
  HOUSE = "House",
}